Move SimpleSqlAsVectorOf
changelog_begin
changelog_end
